corrected two typo bugs in reconfigurationProblem
makeIsMigrated used == instead of != o return test, preventing it from
insatntiating the variable.

extractData used the bad values of destination/source configuration to
set the host/migrationTarget of a VM.

also corrected the test of premigrateTest() because a VM which is
migrated as a solution must be migrating in the destination
configuration.