fixes and refactoring
moved some methods of VariableManager to ConstraintHelper.
NetworkDataBridge does not consider VM couples if a VM is not present.
VM Waiting state are now equal if they are in a couple of the network
view.
corrected VM couple links use when the VM are both waiting, byut not
tested because of choco bug.